@@ -21,11 +21,11 @@ import (
 	"bufio"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strconv"
 	"strings"
+	"time"
 
 	"github.com/elastic/elastic-agent-libs/logp"
 	"github.com/elastic/elastic-agent-system-metrics/metric/system/resolve"
@@ -68,7 +68,7 @@ type PathList struct {
 
 // Flatten combines the V1 and V2 cgroups in cases where we don't need a map with keys
 func (pl PathList) Flatten() []ControllerPath {
-	list := []ControllerPath{}
+	list := make([]ControllerPath, 0, len(pl.V1)+len(pl.V2))
 	for _, v1 := range pl.V1 {
 		list = append(list, v1)
 	}
@@ -230,13 +230,18 @@ func SubsystemMountpoints(rootfs resolve.Resolver, subsystems map[string]struct{
 
 // ProcessCgroupPaths returns the cgroups to which a process belongs and the
 // pathname of the cgroup relative to the mountpoint of the subsystem.
-func (r Reader) ProcessCgroupPaths(pid int) (PathList, error) {
+func (r *Reader) ProcessCgroupPaths(pid int) (PathList, error) {
 	cgroupPath := filepath.Join("proc", strconv.Itoa(pid), "cgroup")
 	cgroup, err := os.Open(r.rootfsMountpoint.ResolveHostFS(cgroupPath))
 	if err != nil {
 		return PathList{}, err //return a blank error so other events can use any file not found errors
 	}
 	defer cgroup.Close()
+
+	version, err := r.CgroupsVersion(pid)
+	if err != nil {
+		return PathList{}, fmt.Errorf("error finding cgroup version for pid %d: %w", pid, err)
+	}
 
 	cPaths := PathList{V1: map[string]ControllerPath{}, V2: map[string]ControllerPath{}}
 	sc := bufio.NewScanner(cgroup)
@@ -258,13 +263,21 @@ func (r Reader) ProcessCgroupPaths(pid int) (PathList, error) {
 		}
 		// cgroup V2
 		// cgroup v2 controllers will always start with this string
-		if strings.Contains(line, "0::/") {
+		if strings.HasPrefix(line, "0::/") {
 			// if you're running inside a container
 			// that's operating with a hybrid cgroups config,
 			// the containerized process won't see the V2 mount
 			// inside /proc/self/mountinfo if docker is using cgroups V1
 			// For this very annoying edge case, revert to the hostfs flag
 			// If it's not set, warn the user that they've hit this.
+
+			// we skip reading paths in case there are cgroups V1 controllers, we are at the cgroup V2 root and the cgroup V2 mount is not available
+			// instead of returning an error because we don't want to break V1 metric collection for misconfigured hybrid systems that have only
+			// a cgroup V2 root but don't have any other controllers. This case happens when cgroup V2 FS is mounted at a special location but not used
+			if version == CgroupsV1 && line == "0::/" && r.cgroupMountpoints.V2Loc == "" {
+				continue
+			}
+
 			controllerPath := filepath.Join(r.cgroupMountpoints.V2Loc, path)
 			if r.cgroupMountpoints.V2Loc == "" && !r.rootfsMountpoint.IsSet() {
 				logp.L().Debugf(`PID %d contains a cgroups V2 path (%s) but no V2 mountpoint was found.
@@ -276,7 +289,25 @@ the container as /sys/fs/cgroup/unified and start the system module with the hos
 				controllerPath = r.rootfsMountpoint.ResolveHostFS(filepath.Join("/sys/fs/cgroup/unified", path))
 			}
 
-			cgpaths, err := ioutil.ReadDir(controllerPath)
+			// Check if there is an entry for controllerPath already cached.
+			r.v2ControllerPathCache.Lock()
+			cacheEntry, ok := r.v2ControllerPathCache.cache[controllerPath]
+			if ok {
+				// If the cached entry for controllerPath is not older than 5 minutes,
+				// return the cached entry.
+				if time.Since(cacheEntry.added) < 5*time.Minute {
+					cPaths.V2 = cacheEntry.pathList.V2
+					r.v2ControllerPathCache.Unlock()
+					continue
+				}
+
+				// Consider the existing entry for controllerPath invalid, as it is
+				// older than 5 minutes.
+				delete(r.v2ControllerPathCache.cache, controllerPath)
+			}
+			r.v2ControllerPathCache.Unlock()
+
+			cgpaths, err := os.ReadDir(controllerPath)
 			if err != nil {
 				return cPaths, fmt.Errorf("error fetching cgroupV2 controllers for cgroup location '%s' and path line '%s': %w", r.cgroupMountpoints.V2Loc, line, err)
 			}
@@ -288,6 +319,12 @@ the container as /sys/fs/cgroup/unified and start the system module with the hos
 					cPaths.V2[controllerName] = ControllerPath{ControllerPath: path, FullPath: controllerPath, IsV2: true}
 				}
 			}
+			r.v2ControllerPathCache.Lock()
+			r.v2ControllerPathCache.cache[controllerPath] = pathListWithTime{
+				added:    time.Now(),
+				pathList: cPaths,
+			}
+			r.v2ControllerPathCache.Unlock()
 			// cgroup v1
 		} else {
 			subsystems := strings.Split(fields[1], ",")
