@@ -16,14 +16,18 @@
 // under the License.
 
 //go:build darwin || freebsd || linux || windows
-// +build darwin freebsd linux windows
 
 package process
 
 import (
+	"bytes"
+	"fmt"
 	"os"
+	"os/exec"
+	"path"
 	"runtime"
 	"sort"
+	"strings"
 	"testing"
 	"time"
 
@@ -38,11 +42,63 @@ import (
 	"github.com/elastic/elastic-agent-system-metrics/metric/system/resolve"
 )
 
+// BenchmarkGetProcess runs a benchmark of the GetProcess method with caching
+// of the command line and environment variables.
+func BenchmarkGetProcess(b *testing.B) {
+	stat, err := initTestResolver()
+	if err != nil {
+		b.Fatalf("Failed init: %s", err)
+	}
+	procs := make(map[int]mapstr.M, 1)
+	pid := os.Getpid()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+
+		process, err := stat.GetOne(pid)
+		if err != nil {
+			continue
+		}
+
+		procs[pid] = process
+	}
+}
+
+func BenchmarkGetTop(b *testing.B) {
+	stat, err := initTestResolver()
+	if err != nil {
+		b.Fatalf("Failed init: %s", err)
+	}
+	procs := make(map[int][]mapstr.M)
+
+	for i := 0; i < b.N; i++ {
+		list, _, err := stat.Get()
+		if err != nil {
+			b.Fatalf("error: %s", err)
+		}
+		procs[i] = list
+	}
+}
+
 func TestGetState(t *testing.T) {
-	// Getpid is really the only way to test this in a cross-platform way
-	state, err := GetPIDState(resolve.NewTestResolver("/"), os.Getpid())
-	require.NoError(t, err)
-	require.Equal(t, Running, state)
+	want := Running
+	pid := os.Getpid()
+	hostfs := resolve.NewTestResolver("/")
+
+	var got PidState
+	var err error
+	test := func() bool {
+		// Getpid is really the only way to test this in a cross-platform way
+		got, err = GetPIDState(hostfs, pid)
+		if err != nil {
+			return false
+		}
+
+		return want == got
+	}
+
+	assert.Eventuallyf(t, test,
+		time.Second, 50*time.Millisecond,
+		"want process state %q, got %q. Last error: %v", want, got, err)
 }
 
 func TestGetOne(t *testing.T) {
@@ -77,8 +133,52 @@ func TestGetOne(t *testing.T) {
 	t.Logf("Proc: %s", procData[0].StringToPrint())
 }
 
+func TestNetworkFetch(t *testing.T) {
+	if runtime.GOOS != "linux" {
+		t.Skip("Network data only available on linux")
+	}
+	testConfig := Stats{
+		Procs:         []string{".*"},
+		Hostfs:        resolve.NewTestResolver("/"),
+		CPUTicks:      false,
+		EnableCgroups: false,
+		EnableNetwork: true,
+	}
+
+	err := testConfig.Init()
+	require.NoError(t, err)
+
+	data, err := testConfig.GetOne(os.Getpid())
+	require.NoError(t, err)
+	networkData, ok := data["network"]
+	require.True(t, ok, "network data not found")
+	require.NotEmpty(t, networkData)
+}
+
+func TestNetworkFilter(t *testing.T) {
+	if runtime.GOOS != "linux" {
+		t.Skip("Network data only available on linux")
+	}
+	testConfig := Stats{
+		Hostfs:         resolve.NewTestResolver("/"),
+		EnableNetwork:  true,
+		NetworkMetrics: []string{"Forwarding"},
+	}
+
+	err := testConfig.Init()
+	require.NoError(t, err)
+
+	data, err := testConfig.GetOne(os.Getpid())
+	require.NoError(t, err)
+
+	_, exists := data.GetValue("network.ip.Forwarding")
+	require.NoError(t, exists, "filter did not preserve key")
+	ipMetrics, _ := data.GetValue("network.ip")
+	require.Equal(t, 1, len(ipMetrics.(map[string]interface{})))
+}
+
 func TestFilter(t *testing.T) {
-	//The logic itself is os-independent, so we'll only test this on the platform least likly to have CI issues
+	// The logic itself is os-independent, so we'll only test this on the platform least likely to have CI issues
 	if runtime.GOOS != "linux" {
 		t.Skip("Run on Linux only")
 	}
@@ -96,7 +196,10 @@ func TestFilter(t *testing.T) {
 
 	procData, _, err := testConfig.Get()
 	assert.NoError(t, err, "GetOne")
-	assert.Equal(t, 2, len(procData))
+	// the total count of processes can either be one or two,
+	// depending on if the highest-mem-usage process and
+	// highest-cpu-usage process are the same.
+	assert.GreaterOrEqual(t, len(procData), 1)
 
 	testZero := Stats{
 		Procs:  []string{".*"},
@@ -179,6 +282,8 @@ func TestGetProcess(t *testing.T) {
 	case "linux":
 		assert.True(t, (len(process.Cwd) > 0))
 	}
+
+	assert.NotEmptyf(t, process.Cmdline, "cmdLine must be present")
 }
 
 func TestMatchProcs(t *testing.T) {
@@ -244,7 +349,7 @@ func TestProcCpuPercentage(t *testing.T) {
 	}
 
 	newState := GetProcCPUPercentage(p1, p2)
-	//GetProcCPUPercentage wil return a number that varies based on the host, due to NumCPU()
+	// GetProcCPUPercentage wil return a number that varies based on the host, due to NumCPU()
 	// So "un-normalize" it, then re-normalized with a constant.
 	cpu := float64(runtime.NumCPU())
 	unNormalized := newState.CPU.Total.Norm.Pct.ValueOr(0) * cpu
@@ -252,43 +357,6 @@ func TestProcCpuPercentage(t *testing.T) {
 
 	assert.EqualValues(t, 0.0721, normalizedTest)
 	assert.EqualValues(t, 3.459, newState.CPU.Total.Pct.ValueOr(0))
-}
-
-// BenchmarkGetProcess runs a benchmark of the GetProcess method with caching
-// of the command line and environment variables.
-func BenchmarkGetProcess(b *testing.B) {
-	stat, err := initTestResolver()
-	if err != nil {
-		b.Fatalf("Failed init: %s", err)
-	}
-	procs := make(map[int]mapstr.M, 1)
-	pid := os.Getpid()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-
-		process, err := stat.GetOne(pid)
-		if err != nil {
-			continue
-		}
-
-		procs[pid] = process
-	}
-}
-
-func BenchmarkGetTop(b *testing.B) {
-	stat, err := initTestResolver()
-	if err != nil {
-		b.Fatalf("Failed init: %s", err)
-	}
-	procs := make(map[int][]mapstr.M)
-
-	for i := 0; i < b.N; i++ {
-		list, _, err := stat.Get()
-		if err != nil {
-			b.Fatalf("error: %s", err)
-		}
-		procs[i] = list
-	}
 }
 
 func TestIncludeTopProcesses(t *testing.T) {
@@ -511,6 +579,58 @@ func TestIncludeTopProcesses(t *testing.T) {
 	}
 }
 
+// runThreads run the threads binary for the current GOOS.
+//
+//go:generate docker run --rm -v ./testdata:/app --entrypoint g++ docker.elastic.co/beats-dev/golang-crossbuild:1.21.0-main -pthread -std=c++11 -o /app/threads /app/threads.cpp
+//go:generate docker run --rm -v ./testdata:/app --entrypoint o64-clang++ docker.elastic.co/beats-dev/golang-crossbuild:1.21.0-darwin -pthread -std=c++11 -o /app/threads-darwin /app/threads.cpp
+//go:generate docker run --rm -v ./testdata:/app --entrypoint x86_64-w64-mingw32-g++-posix docker.elastic.co/beats-dev/golang-crossbuild:1.21.0-main -pthread -std=c++11 -o /app/threads.exe /app/threads.cpp
+func runThreads(t *testing.T) *exec.Cmd { //nolint: deadcode,structcheck,unused // needed by other platforms
+	t.Helper()
+
+	supportedPlatforms := []string{"linux/amd64", "darwin/amd64", "windows/amd64"}
+
+	platform := fmt.Sprintf("%s/%s", runtime.GOOS, runtime.GOARCH)
+	if !sliceContains(supportedPlatforms, platform) {
+		t.Skipf("not supported for %s/%s. Supported patforms: %v",
+			runtime.GOOS, runtime.GOARCH, supportedPlatforms)
+	}
+
+	threads := path.Join("testdata", "threads")
+
+	switch runtime.GOOS {
+	case "linux":
+		// nothing to do
+	case "darwin":
+		threads += "-darwin"
+	case "windows":
+		threads += ".exe"
+	}
+
+	var b bytes.Buffer
+	cmd := exec.Command(threads)
+	cmd.Stdout = &b
+	cmd.Stderr = &b
+
+	err := cmd.Start()
+	require.NoErrorf(t, err, "failed to start %q", threads)
+
+	var log string
+	require.Eventually(t,
+		func() bool {
+			if cmd.ProcessState != nil {
+				t.Fatalf("Process exited with error: '%s'", cmd.ProcessState.String())
+			}
+			line := b.String()
+			log += line
+			return strings.Contains(log, "running")
+		},
+		time.Second, 50*time.Millisecond,
+		"could not determine if %q is running. Output: %q",
+		threads, log)
+
+	return cmd
+}
+
 func initTestResolver() (Stats, error) {
 	err := logp.DevelopmentSetup()
 	if err != nil {
@@ -535,4 +655,14 @@ func initTestResolver() (Stats, error) {
 	}
 	err = testConfig.Init()
 	return testConfig, err
+}
+
+func sliceContains(s []string, e string) bool { //nolint: deadcode,structcheck,unused // needed by other platforms
+	for _, v := range s {
+		if e == v {
+			return true
+		}
+	}
+
+	return false
 }
