@@ -23,17 +23,19 @@ import (
 	"github.com/elastic/elastic-agent-libs/mapstr"
 	"github.com/elastic/elastic-agent-libs/opt"
 	"github.com/elastic/elastic-agent-system-metrics/metric/system/cgroup"
+	sysinfotypes "github.com/elastic/go-sysinfo/types"
 )
 
 // ProcState is the main struct for process information and metrics.
 type ProcState struct {
 	// Basic Process data
-	Name     string   `struct:"name,omitempty"`
-	State    PidState `struct:"state,omitempty"`
-	Username string   `struct:"username,omitempty"`
-	Pid      opt.Int  `struct:"pid,omitempty"`
-	Ppid     opt.Int  `struct:"ppid,omitempty"`
-	Pgid     opt.Int  `struct:"pgid,omitempty"`
+	Name       string   `struct:"name,omitempty"`
+	State      PidState `struct:"state,omitempty"`
+	Username   string   `struct:"username,omitempty"`
+	Pid        opt.Int  `struct:"pid,omitempty"`
+	Ppid       opt.Int  `struct:"ppid,omitempty"`
+	Pgid       opt.Int  `struct:"pgid,omitempty"`
+	NumThreads opt.Int  `struct:"num_threads,omitempty"`
 
 	// Extended Process Data
 	Args    []string `struct:"args,omitempty"`
@@ -43,9 +45,11 @@ type ProcState struct {
 	Env     mapstr.M `struct:"env,omitempty"`
 
 	// Resource Metrics
-	Memory ProcMemInfo `struct:"memory,omitempty"`
-	CPU    ProcCPUInfo `struct:"cpu,omitempty"`
-	FD     ProcFDInfo  `struct:"fd,omitempty"`
+	Memory  ProcMemInfo                       `struct:"memory,omitempty"`
+	CPU     ProcCPUInfo                       `struct:"cpu,omitempty"`
+	FD      ProcFDInfo                        `struct:"fd,omitempty"`
+	Network *sysinfotypes.NetworkCountersInfo `struct:"-,omitempty"`
+	IO      ProcIOInfo                        `struct:"io,omitempty"`
 
 	// cgroups
 	Cgroup cgroup.CGStats `struct:"cgroup,omitempty"`
@@ -74,6 +78,24 @@ type CPUTotal struct {
 	Ticks opt.Uint   `struct:"ticks,omitempty"`
 	Pct   opt.Float  `struct:"pct,omitempty"`
 	Norm  opt.PctOpt `struct:"norm,omitempty"`
+}
+
+// ProcIOInfo is the struct for I/O counters from /proc/[pid]/io
+type ProcIOInfo struct {
+	// ReadChar is bytes read from the system, as passed from read() and similar syscalls
+	ReadChar opt.Uint `struct:"read_char,omitempty"`
+	// WriteChar is bytes written to the system, as passed to various syscalls
+	WriteChar opt.Uint `struct:"write_char,omitempty"`
+	//ReadSyscalls counts the number of read operations
+	ReadSyscalls opt.Uint `struct:"read_ops,omitempty"`
+	//WriteSyscalls counts the number of write operations
+	WriteSyscalls opt.Uint `struct:"write_ops,omitempty"`
+	// ReadBytes is the count of bytes that were actually fetched from the storage layer
+	ReadBytes opt.Uint `struct:"read_bytes,omitempty"`
+	// WriteBytes is the count of bytes that were actually written to the storage layer
+	WriteBytes opt.Uint `struct:"write_bytes,omitempty"`
+	// the number of bytes which this process caused to not happen, by truncating pagecache
+	CancelledWriteBytes opt.Uint `struct:"cancelled_write_bytes,omitempty"`
 }
 
 // ProcMemInfo is the struct for cpu.memory metrics
