@@ -18,17 +18,17 @@
 package process
 
 import (
+	"errors"
 	"fmt"
 	"path/filepath"
 	"syscall"
+	"unsafe"
+
+	xsyswindows "golang.org/x/sys/windows"
 
 	"github.com/elastic/elastic-agent-libs/opt"
 	"github.com/elastic/elastic-agent-system-metrics/metric/system/resolve"
 	"github.com/elastic/gosigar/sys/windows"
-)
-
-var (
-	processQueryLimitedInfoAccess = windows.PROCESS_QUERY_LIMITED_INFORMATION
 )
 
 // FetchPids returns a map and array of pids
@@ -38,9 +38,11 @@ func (procStats *Stats) FetchPids() (ProcsMap, []ProcState, error) {
 		return nil, nil, fmt.Errorf("EnumProcesses failed: %w", err)
 	}
 
-	procMap := make(ProcsMap, 0)
-	var plist []ProcState
-	// This is probably the only implementation that doesn't benefit from our little fillPid callback system. We'll need to iterate over everything manually.
+	procMap := make(ProcsMap, len(pids))
+	plist := make([]ProcState, 0, len(pids))
+	// This is probably the only implementation that doesn't benefit from our
+	// little fillPid callback system. We'll need to iterate over everything
+	// manually.
 	for _, pid := range pids {
 		procMap, plist = procStats.pidIter(int(pid), procMap, plist)
 	}
@@ -50,23 +52,70 @@ func (procStats *Stats) FetchPids() (ProcsMap, []ProcState, error) {
 
 // GetInfoForPid returns basic info for the process
 func GetInfoForPid(_ resolve.Resolver, pid int) (ProcState, error) {
-	state := ProcState{}
+	var err error
+	var errs []error
+	state := ProcState{Pid: opt.IntWith(pid)}
 
 	name, err := getProcName(pid)
 	if err != nil {
-		return state, fmt.Errorf("error fetching name: %w", err)
+		errs = append(errs, fmt.Errorf("error fetching name: %w", err))
+	} else {
+		state.Name = name
 	}
-	state.Name = name
-	state.Pid = opt.IntWith(pid)
 
 	// system/process doesn't need this here, but system/process_summary does.
 	status, err := getPidStatus(pid)
 	if err != nil {
-		return state, fmt.Errorf("error fetching status: %w", err)
+		errs = append(errs, fmt.Errorf("error fetching status: %w", err))
+	} else {
+		state.State = status
 	}
-	state.State = status
+
+	if err := errors.Join(errs...); err != nil {
+		return state, fmt.Errorf("could not get all information for PID %d: %w",
+			pid, err)
+	}
 
 	return state, nil
+}
+
+func FetchNumThreads(pid int) (int, error) {
+	targetProcessHandle, err := syscall.OpenProcess(
+		xsyswindows.PROCESS_QUERY_INFORMATION,
+		false,
+		uint32(pid))
+	if err != nil {
+		return 0, fmt.Errorf("OpenProcess failed for PID %d: %w", pid, err)
+	}
+	defer syscall.CloseHandle(targetProcessHandle)
+
+	currentProcessHandle, err := syscall.GetCurrentProcess()
+	if err != nil {
+		return 0, fmt.Errorf("GetCurrentProcess failed: %w", err)
+	}
+	// The pseudo handle need not be closed when it is no longer
+	// needed, calling CloseHandle has no effect.  Adding here to
+	// remind us to close any handles we open.
+	defer syscall.CloseHandle(currentProcessHandle)
+
+	var snapshotHandle syscall.Handle
+	err = PssCaptureSnapshot(targetProcessHandle, PSSCaptureThreads, 0, &snapshotHandle)
+	if err != nil {
+		return 0, fmt.Errorf("PssCaptureSnapshot failed: %w", err)
+	}
+
+	info := PssThreadInformation{}
+	buffSize := unsafe.Sizeof(info)
+	queryErr := PssQuerySnapshot(snapshotHandle, PssQueryThreadInformation, &info, uint32(buffSize))
+	freeErr := PssFreeSnapshot(currentProcessHandle, snapshotHandle)
+	if queryErr != nil || freeErr != nil {
+		//Join discards any nil errors
+		return 0, errors.Join(
+			fmt.Errorf("PssQuerySnapshot failed: %w", queryErr),
+			fmt.Errorf("PssFreeSnapshot failed: %w", freeErr))
+	}
+
+	return int(info.ThreadsCaptured), nil
 }
 
 // FillPidMetrics is the windows implementation
@@ -98,17 +147,34 @@ func FillPidMetrics(_ resolve.Resolver, pid int, state ProcState, _ func(string)
 
 	state.CPU.StartTime = unixTimeMsToTime(startTime)
 
+	return state, nil
+}
+
+// FillMetricsRequiringMoreAccess
+// All calls that need more access rights than
+// windows.PROCESS_QUERY_LIMITED_INFORMATION
+func FillMetricsRequiringMoreAccess(pid int, state ProcState) (ProcState, error) {
 	argList, err := getProcArgs(pid)
 	if err != nil {
-		return state, fmt.Errorf("error fetching process args: %w", err)
+		return state, fmt.Errorf("error fetching process args: %w", NonFatalErr{Err: err})
 	}
 	state.Args = argList
+
+	if numThreads, err := FetchNumThreads(pid); err != nil {
+		return state, fmt.Errorf("error fetching num threads: %w", NonFatalErr{Err: err})
+	} else {
+		state.NumThreads = opt.IntWith(numThreads)
+	}
+
 	return state, nil
 }
 
 func getProcArgs(pid int) ([]string, error) {
-
-	handle, err := syscall.OpenProcess(processQueryLimitedInfoAccess|windows.PROCESS_VM_READ, false, uint32(pid))
+	handle, err := syscall.OpenProcess(
+		windows.PROCESS_QUERY_LIMITED_INFORMATION|
+			windows.PROCESS_VM_READ,
+		false,
+		uint32(pid))
 	if err != nil {
 		return nil, fmt.Errorf("OpenProcess failed: %w", err)
 	}
@@ -137,7 +203,7 @@ func getProcArgs(pid int) ([]string, error) {
 }
 
 func getProcTimes(pid int) (uint64, uint64, uint64, error) {
-	handle, err := syscall.OpenProcess(processQueryLimitedInfoAccess, false, uint32(pid))
+	handle, err := syscall.OpenProcess(windows.PROCESS_QUERY_LIMITED_INFORMATION, false, uint32(pid))
 	if err != nil {
 		return 0, 0, 0, fmt.Errorf("OpenProcess failed for pid=%v: %w", pid, err)
 	}
@@ -154,8 +220,16 @@ func getProcTimes(pid int) (uint64, uint64, uint64, error) {
 	return uint64(windows.FiletimeToDuration(&cpu.UserTime).Nanoseconds() / 1e6), uint64(windows.FiletimeToDuration(&cpu.KernelTime).Nanoseconds() / 1e6), uint64(cpu.CreationTime.Nanoseconds() / 1e6), nil
 }
 
+// procMem gets the memory usage for the given PID.
+// The current implementation calls
+// GetProcessMemoryInfo (https://learn.microsoft.com/en-us/windows/win32/api/psapi/nf-psapi-getprocessmemoryinfo)
+// We only need `PROCESS_QUERY_LIMITED_INFORMATION` because we do not support
+// Windows Server 2003 or Windows XP
 func procMem(pid int) (uint64, uint64, error) {
-	handle, err := syscall.OpenProcess(processQueryLimitedInfoAccess|windows.PROCESS_VM_READ, false, uint32(pid))
+	handle, err := syscall.OpenProcess(
+		windows.PROCESS_QUERY_LIMITED_INFORMATION,
+		false,
+		uint32(pid))
 	if err != nil {
 		return 0, 0, fmt.Errorf("OpenProcess failed for pid=%v: %w", pid, err)
 	}
@@ -172,7 +246,7 @@ func procMem(pid int) (uint64, uint64, error) {
 
 // getProcName returns the process name associated with the PID.
 func getProcName(pid int) (string, error) {
-	handle, err := syscall.OpenProcess(processQueryLimitedInfoAccess, false, uint32(pid))
+	handle, err := syscall.OpenProcess(windows.PROCESS_QUERY_LIMITED_INFORMATION, false, uint32(pid))
 	if err != nil {
 		return "", fmt.Errorf("OpenProcess failed for pid=%v: %w", pid, err)
 	}
@@ -190,7 +264,7 @@ func getProcName(pid int) (string, error) {
 
 // getProcStatus returns the status of a process.
 func getPidStatus(pid int) (PidState, error) {
-	handle, err := syscall.OpenProcess(processQueryLimitedInfoAccess, false, uint32(pid))
+	handle, err := syscall.OpenProcess(windows.PROCESS_QUERY_LIMITED_INFORMATION, false, uint32(pid))
 	if err != nil {
 		return Unknown, fmt.Errorf("OpenProcess failed for pid=%v: %w", pid, err)
 	}
@@ -204,7 +278,7 @@ func getPidStatus(pid int) (PidState, error) {
 		return Unknown, fmt.Errorf("GetExitCodeProcess failed for pid=%v: %w", pid, err)
 	}
 
-	if exitCode == 259 { //still active
+	if exitCode == 259 { // still active
 		return Running, nil
 	}
 	return Sleeping, nil
@@ -212,7 +286,7 @@ func getPidStatus(pid int) (PidState, error) {
 
 // getParentPid returns the parent process ID of a process.
 func getParentPid(pid int) (int, error) {
-	handle, err := syscall.OpenProcess(processQueryLimitedInfoAccess, false, uint32(pid))
+	handle, err := syscall.OpenProcess(windows.PROCESS_QUERY_LIMITED_INFORMATION, false, uint32(pid))
 	if err != nil {
 		return 0, fmt.Errorf("OpenProcess failed for pid=%v: %w", pid, err)
 	}
@@ -229,7 +303,7 @@ func getParentPid(pid int) (int, error) {
 }
 
 func getProcCredName(pid int) (string, error) {
-	handle, err := syscall.OpenProcess(syscall.PROCESS_QUERY_INFORMATION, false, uint32(pid))
+	handle, err := syscall.OpenProcess(windows.PROCESS_QUERY_LIMITED_INFORMATION, false, uint32(pid))
 	if err != nil {
 		return "", fmt.Errorf("OpenProcess failed for pid=%v: %w", pid, err)
 	}
